+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/MainConfiguration.php
@@ -319,7 +319,7 @@ class MainConfiguration implements ConfigurationInterface
                             'memory' => [
                                 'users' => [
                                     'foo' => ['password' => 'foo', 'roles' => 'ROLE_USER'],
-                                    'bar' => ['password' => 'bar', 'roles' => '[ROLE_USER, ROLE_ADMIN]'],
+                                    'bar' => ['password' => 'bar', 'roles' => ['ROLE_USER', 'ROLE_ADMIN'], 'extra_fields' => ['name' => 'John', 'age' => 77]],
                                 ],
                             ],
                         ],